Crate piecetable

Implementation of a piece table based on a vector. A piece table provides efficient methods for inserting and removing elements sequentially, intended for use as the underlying data structure in a text editor.

The piece table stores a read only reference to a source (if one is provided) and stores inserted elements in an append-only vector. A table of pieces pointing either to the source or add-buffer is maintained, and these pieces are manipulated when inserting and removing text. Asymptotics in the following are based on p, the number of pieces, where p should be strictly smaller than the number of elements when used as intended.

Structs

Iter
Struct for iterating the elements of a PieceTable.

PieceTable
The PieceTable type with all relevant methods.

Range
Struct for iterating a range of elements in a PieceTable.