Struct piecetable::PieceTable

pub struct PieceTable<'a, T: 'a> {
    // some fields omitted
}

The PieceTable type with all relevant methods.

Methods

impl<'a, T: 'a> PieceTable<'a, T>

fn new() -> PieceTable<'a, T>

Construct a new PieceTable

Example

use piecetable::PieceTable;
let table: PieceTable<char> = PieceTable::new();

fn with_capacity(data_capacity: usize, piece_capacity: usize) -> PieceTable<'a, T>

Constructs a new, empty PieceTable<T> with the specified capacity for elements and pieces. Sequential insertion of data_capacity elements will be possible without reallocation. Scattered operations results in pieces being added; the created piece table will be able to store piece_capacity of these before reallocating.

fn src(self, src: &'a [T]) -> PieceTable<'a, T>

Assign a read-only source to an existing PieceTable.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..100).collect();
let table = PieceTable::new().src(&src);

fn len(&self) -> usize

The number of elements stored in the piece table.

fn is_empty(&self) -> bool

fn capacity_data(&self) -> usize

Returns the number of elements sequentially inserted the piece table can hold without reallocating.

fn capacity_pieces(&self) -> usize

Returns the number of pieces (created by scattered operations), the piece table can hold without reallocating.

fn reserve_data(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted. The collection may reserve more space to avoid frequent reallocations.

fn reserve_piece(&mut self, additional: usize)

Reserves capacity for at least additional more pieces to be created. The collection may reserve more space to avoid frequent reallocations.

fn clear(&mut self)

Clears the piece table, removing all elements. Also removes reference to any given src.

fn iter(&'a self) -> Iter<'a, T>

Return an iterator over all elements of the PieceTable.

Advancing the iterator takes constant time.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..101).collect();
let table = PieceTable::new().src(&src);
assert_eq!(5050, table.iter().fold(0, |acc, &x| acc + x));

fn range(&'a self, min: Bound<usize>, max: Bound<usize>) -> Range<'a, T>

Return an iterator over the bound range in the PieceTable. Constructing the iterator takes O(p) time, but consuming it is constant time per element.

Example

#![feature(collections_bound)]
use piecetable::PieceTable;
use std::collections::Bound::*;
let src: Vec<i32> = (0..100).collect();
let table = PieceTable::new().src(&src);
assert_eq!(vec![&55, &56, &57], table.range(Included(55), Excluded(58)).collect::<Vec<&i32>>());

fn insert(&mut self, idx: usize, item: T)

Insert an element at idx.

O(p) time, but sequential inserts afterwards take O(1)time.

Panics

Panics if not idx <= len.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..100).collect();
let mut table = PieceTable::new().src(&src);
table.insert(11, 42); // takes `O(p)` time
table.insert(12, 42); // constant time
table.insert(13, 42); // constant time
table.insert(27, 42); // `O(p)` time
table.insert(28, 42); // constant time

fn remove(&mut self, idx: usize)

Remove the element at the given index.

O(p) operation initially, but removing sequentially backwards afterwards is O(1).

Panics

Panics if not idx < len.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..10).collect();
let mut table = PieceTable::new().src(&src);
table.remove(5); // `O(p)`
table.remove(4); // `O(1)`
table.remove(3); // `O(1)`
table.remove(6); // `O(p)`
assert_eq!(vec![&0, &1, &2, &6, &7, &8], table.iter().collect::<Vec<&i32>>());

fn push(&mut self, value: T)

Appends an element to the back, efficiently and in constant time.