Struct piecetable::PieceTable

pub struct PieceTable<'a, T: 'a> {
    // some fields omitted
}

The PieceTable type with all relevant methods.

Methods

impl<'a, T: 'a> PieceTable<'a, T>

fn new() -> PieceTable<'a, T>

Construct a new PieceTable

Example

use piecetable::PieceTable;
let table: PieceTable<char> = PieceTable::new();

fn src(self, src: &'a [T]) -> PieceTable<'a, T>

Assign a read-only source to an existing PieceTable.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..100).collect();
let table = PieceTable::new().src(&src);

fn iter(&'a self) -> Iter<'a, T>

Return an iterator over all elements of the PieceTable.

Advancing the iterator takes constant time.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..101).collect();
let table = PieceTable::new().src(&src);
assert_eq!(5050, table.iter().fold(0, |acc, &x| acc + x));

fn range(&'a self, from: usize, to: usize) -> Range<'a, T>

Return an iterator over the range [from, to) in the PieceTable. Construction the iterator takes O(p) time, but consuming it is constant time per element.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..100).collect();
let table = PieceTable::new().src(&src);
assert_eq!(vec![&55, &56, &57], table.range(55, 58).collect::<Vec<&i32>>());

fn insert(&mut self, idx: usize, item: T)

Insert an element at idx.

O(p) time, but sequential inserts afterwards take O(1)time.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..100).collect();
let mut table = PieceTable::new().src(&src);
table.insert(11, 42); // takes `O(p)` time
table.insert(12, 42); // constant time
table.insert(13, 42); // constant time
table.insert(27, 42); // `O(p)` time
table.insert(28, 42); // constant time

fn remove(&mut self, idx: usize)

Remove the element at the given index.

O(p) operation, but removing just inserted indices sequentially is O(1) and most remove operations are cheap.

Example

use piecetable::PieceTable;
let src: Vec<i32> = (0..10).collect();
let mut table = PieceTable::new().src(&src);
table.remove(5);
table.remove(6);
assert_eq!(vec![&0, &1, &2, &3, &4, &6, &8, &9], table.iter().collect::<Vec<&i32>>());